% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMixHMMR.R
\docType{class}
\name{StatMixHMMR-class}
\alias{StatMixHMMR-class}
\alias{StatMixHMMR}
\title{A Reference Class which contains statistics of a MixHMMR model.}
\description{
StatMixHMMR contains all the statistics associated to a
\link[=ParamMixHMMR]{MixHMMR} model.
}
\section{Fields}{

\describe{
\item{\code{tau_ik}}{Matrix of size \eqn{(n, K)} giving the posterior probabilities
that the curve \eqn{Y_{i}} originates from the \eqn{k}-th HMM model.}

\item{\code{gamma_ikjr}}{Array of size \eqn{(nm, R, K)} giving the posterior
probabilities that the observation \eqn{Y_{ij}} originates from the
\eqn{r}-th regime of the \eqn{k}-th HMM model.}

\item{\code{loglik}}{Numeric. Log-likelihood of the MixHMMR model.}

\item{\code{stored_loglik}}{List. Stored values of the log-likelihood at each
iteration of the EM algorithm.}

\item{\code{cputime}}{Numeric. Average computing time of a EM algorithm run.}

\item{\code{klas}}{Row matrix of the labels issued from \code{tau_ik}. Its elements are
\eqn{klas(i) = k}, \eqn{i = 1,\dots,n}.}

\item{\code{z_ik}}{Hard segmentation logical matrix of dimension \eqn{(n, K)}
obtained by the Maximum a posteriori (MAP) rule: \eqn{z\_ik = 1 \
  \textrm{if} \ z\_ik = \textrm{arg} \ \textrm{max}_{s} \ P(z_{is} = 1 |
  \boldsymbol{Y_{i}}; \boldsymbol{\Psi}) = tau\_tk;\ 0 \
  \textrm{otherwise}}{z_ik = 1 if z_ik = arg max_s P(z_{is} = 1 | Y_{i};
  \Psi) = tau_tk; 0 otherwise}, \eqn{k = 1,\dots,K}.}

\item{\code{smoothed}}{Matrix of size \eqn{(m, K)} giving the estimated mean series.
The k-th column gives the estimated mean series of cluster k.}

\item{\code{mean_curve}}{To define.}

\item{\code{BIC}}{Numeric. Value of BIC (Bayesian Information Criterion).}

\item{\code{AIC}}{Numeric. Value of AIC (Akaike Information Criterion).}

\item{\code{ICL1}}{Numeric. Value of ICL (Integrated Completed Likelihood
Criterion).}

\item{\code{log_w_k_fyi}}{Private. Only defined for calculations.}

\item{\code{exp_num_trans}}{Private. Only defined for calculations.}

\item{\code{exp_num_trans_from_l}}{Private. Only defined for calculations.}
}}


\seealso{
\link{ParamMixHMMR}
}
